+++ include/global_variables.h
@@ -3,3 +3,6 @@
 
 extern PlayState* gPlayState;
 extern s8 current_room;
+
+extern RoomContext *gRoomCtx;
+extern s8 gRoomStatus;